wrap "-e" arguments containing spaces in shell
When passing the terminal as a string to be executed, some programs
expect the flag "-e" to accept the command as a single space
deliminated string.

This is implemented by checking if only a single argument follows "-e"
and if this contains spaces, then wrapping the argument in "sh -c".